wc18-back/wc18-domain/src/main/java/com/github/mjeanroy/wc18/domain/dao/StageDao.java
/**
 * The MIT License (MIT)
 *
 * Copyright (c) 2018 Mickael Jeanroy
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */

package com.github.mjeanroy.wc18.domain.dao;

import com.github.mjeanroy.wc18.domain.models.Stage;
import org.springframework.stereotype.Component;

import java.util.Arrays;
import java.util.Map;
import java.util.Optional;
import java.util.stream.Collectors;

import static java.util.Arrays.asList;
import static java.util.function.Function.identity;

@Component
public class StageDao {

	/**
	 * The list of stages.
	 */
	private final Map<String, Stage> stages;

	public StageDao() {
		this.stages = Arrays.stream(Stage.values()).collect(Collectors.toMap(Stage::name, identity()));
	}

	/**
	 * Get all stages.
	 *
	 * @return Stages.
	 */
	public Iterable<Stage> findAll() {
		return asList(Stage.values());
	}

	/**
	 * Find stage by its identifier.
	 *
	 * @param id Stage identifier.
	 * @return The stage.
	 */
	public Optional<Stage> findOne(String id) {
		return Optional.ofNullable(stages.getOrDefault(id, null));
	}
}
